refactor(timeframe): moved duration filter into a separate component:
