+++ galasa-ui/src/components/test-runs/timeframe/DurationFilter.tsx
@@ -1,0 +1,65 @@
+/*
+ * Copyright contributors to the Galasa project
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+'use client';
+
+import { TimeFrameValues } from '@/utils/interfaces';
+import { FormGroup, NumberInput } from '@carbon/react';
+import { useTranslations } from 'next-intl';
+import styles from '@/styles/TimeFrameContent.module.css';
+
+const MAX_DAYS = 90;
+const MAX_HOURS = 23;
+const MAX_MINUTES = 59;
+
+export default function DurationFilter({
+  values,
+  handleValueChange,
+}: {
+  values: TimeFrameValues;
+  handleValueChange: (field: keyof TimeFrameValues, value: any) => void;
+}) {
+  const translations = useTranslations('DurationFilter');
+
+  return (
+    <FormGroup legendText={translations('duration')} className={styles.durationFilterContainer}>
+      <div className={styles.durationInputsContainer} key={values.toDate?.getTime() || 0}>
+        <NumberInput
+          id="duration-days"
+          label={translations('days')}
+          min={0}
+          max={MAX_DAYS}
+          value={values.durationDays}
+          onChange={(
+            _: React.ChangeEvent<HTMLInputElement>,
+            { value }: { value: number | string }
+          ) => handleValueChange('durationDays', value)}
+        />
+        <NumberInput
+          id="duration-hours"
+          label={translations('hours')}
+          min={0}
+          max={MAX_HOURS}
+          value={values.durationHours}
+          onChange={(
+            _: React.ChangeEvent<HTMLInputElement>,
+            { value }: { value: number | string }
+          ) => handleValueChange('durationHours', value)}
+        />
+        <NumberInput
+          id="duration-minutes"
+          label={translations('minutes')}
+          min={0}
+          max={MAX_MINUTES}
+          value={values.durationMinutes}
+          onChange={(
+            _: React.ChangeEvent<HTMLInputElement>,
+            { value }: { value: number | string }
+          ) => handleValueChange('durationMinutes', value)}
+        />
+      </div>
+    </FormGroup>
+  );
+}
