+++ galasa-ui/src/components/test-runs/timeframe/TimeFrameFilter.tsx
@@ -5,15 +5,10 @@
  */
 'use client';
 
-import styles from '@/styles/TestRunsPage.module.css';
+import styles from '@/styles/TimeFrameContent.module.css';
 import { TimeFrameValues } from '@/utils/interfaces';
-import { FormGroup, NumberInput } from '@carbon/react';
 import DateTimePicker from './DateTimePicker';
 import { useTranslations } from 'next-intl';
-
-const MAX_DAYS = 90;
-const MAX_HOURS = 23;
-const MAX_MINUTES = 59;
 
 export default function TimeFrameFilter({
   values,
@@ -24,7 +19,7 @@ export default function TimeFrameFilter({
 }) {
   const translations = useTranslations('TimeFrameFilter');
   return (
-    <div className={styles.TimeFrameFilterContainer}>
+    <div className={styles.timeFrameFilterContainer}>
       <DateTimePicker
         legend={translations('from')}
         date={values.fromDate}
@@ -34,44 +29,6 @@ export default function TimeFrameFilter({
         onTimeChange={(time) => handleValueChange('fromTime', time)}
         onAmPmChange={(amPm) => handleValueChange('fromAmPm', amPm)}
       />
-      <FormGroup legendText={translations('duration')} className={styles.TimeFrameFilterItem}>
-        <div className={styles.DurationInputsContainer} key={values.toDate?.getTime() || 0}>
-          <NumberInput
-            id="duration-days"
-            label={translations('days')}
-            min={0}
-            max={MAX_DAYS}
-            value={values.durationDays}
-            onChange={(
-              _: React.ChangeEvent<HTMLInputElement>,
-              { value }: { value: number | string }
-            ) => handleValueChange('durationDays', value)}
-          />
-          <NumberInput
-            id="duration-hours"
-            label={translations('hours')}
-            min={0}
-            max={MAX_HOURS}
-            value={values.durationHours}
-            onChange={(
-              _: React.ChangeEvent<HTMLInputElement>,
-              { value }: { value: number | string }
-            ) => handleValueChange('durationHours', value)}
-          />
-          <NumberInput
-            id="duration-minutes"
-            label={translations('minutes')}
-            min={0}
-            max={MAX_MINUTES}
-            value={values.durationMinutes}
-            onChange={(
-              _: React.ChangeEvent<HTMLInputElement>,
-              { value }: { value: number | string }
-            ) => handleValueChange('durationMinutes', value)}
-          />
-        </div>
-      </FormGroup>
-
       <DateTimePicker
         legend={translations('to')}
         date={values.toDate}
