+++ galasa-ui/src/styles/TestRunsPage.module.css
@@ -13,7 +13,6 @@
   font-size: 1.2rem;
 }
 
-.timeFrameText,
 .titleText {
   font-size: 1rem;
   margin: 0.5rem 0.5rem;
@@ -161,19 +160,6 @@
   }
 }
 
-.TimeFrameFilterContainer {
-  display: flex;
-  justify-content: space-between;
-  gap: 2rem;
-  width: 85%;
-  padding-top: 1.5rem;
-}
-
-.DurationInputsContainer {
-  display: flex;
-  gap: 1rem;
-}
-
 .notification {
   position: fixed;
   bottom: 5.5rem;
@@ -181,16 +167,6 @@
   z-index: 9999;
 }
 
-@media (max-width: 1055px) {
-  .TimeFrameFilterContainer {
-    flex-direction: column;
-  }
-
-  .DurationInputsContainer {
-    flex-direction: column;
-    gap: 0.5rem;
-  }
-}
 
 /* Styles of the Table Design tab */
