+++ galasa-ui/src/styles/TimeFrameContent.module.css
@@ -1,0 +1,48 @@
+/*
+ * Copyright contributors to the Galasa project
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+.timeFrameText,
+.titleText {
+  font-size: 1rem;
+  margin: 0.5rem 0.5rem;
+}
+
+.timeFrameFilterContainer {
+  display: flex;
+  gap: 10rem;
+  width: 85%;
+  padding-top: 1.5rem;
+}
+
+.durationFilterContainer {
+  display: flex;
+  flex-direction: column;
+  max-width: 600px;
+  min-width: 200px;
+}
+
+.durationInputsContainer {
+  display: flex;
+  gap: 1rem;
+}
+
+.notification {
+  position: fixed;
+  bottom: 5.5rem;
+  right: 1rem;
+  z-index: 9999;
+}
+
+@media (max-width: 1055px) {
+  .TimeFrameFilterContainer {
+    flex-direction: column;
+  }
+
+  .DurationInputsContainer {
+    flex-direction: column;
+    gap: 0.5rem;
+  }
+}
